questão 1 e 2 atividade

+++ games/views.py
@@ -4,6 +4,9 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from games.models import Game
 from games.serializers import GameSerializer
+from datetime import datetime
+import pytz
+
 
 @api_view(['GET','POST'])
 def game_list(request):
@@ -38,9 +41,24 @@ def game_detail(request, pk):
     elif request.method == "PUT":
         game_serializer = GameSerializer(game, data=request.data)
         if game_serializer.is_valid():
-            game_serializer.save()
-            return Response(game_serializer.data)
+            try:
+                games = Game.objects.filter(name=request.data['name'])         
+                if not game in games:
+                    return Response({"name":"game already exists in the database"}, status=status.HTTP_400_BAD_REQUEST)
+                else:
+                    game_serializer.save()
+                    return Response(game_serializer.data)
+            except:
+                game_serializer.save()
+                return Response(game_serializer.data)
         return Response(game_serializer.errors,status=status.HTTP_400_BAD_REQUEST)
     elif request.method == "DELETE":
-        game.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        utc=pytz.UTC
+        date2 = utc.localize(datetime.now())
+        print(date2)
+        # date1 = utc.localize(game.release_date)
+        # print(date1)
+        if game.release_date > date2:        
+            game.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response({"date_release":"Game has been released"},status=status.HTTP_401_UNAUTHORIZED)
